requirementa.txt
transformers
datasets
peft
bitsandbytes
accelerate
llama-cpp-python
